Modelling Performance Indicators

MPI_bias: MPI for bias MPI_corr_time: temporal MPI for correlation MPI_sdev_time: temporal MPI for standard deviation MPI_corr_space: spatial MPI for correlation MPI_sdev_space: spatial MPI for standard deviation MPI_perc: MPI for high percentile values

MPI_bias(obs, mod, pollutant, beta = 2, ...)

MPI_corr_time(obs, mod, pollutant, beta = 2, ...)

MPI_sdev_time(obs, mod, pollutant, beta = 2, ...)

MPI_corr_space(obs, mod, pollutant, beta = 2, ...)

MPI_sdev_space(obs, mod, pollutant, beta = 2, ...)

MPI_perc(obs, mod, pollutant, beta = 2, ...)

Arguments

obs
numeric vector of observed values (yearly averages in MPI_*_space)

mod
numeric vector of modelled values (yearly averages in MPI_*_space)

pollutant
one of "NO2", "O3", "PM10", "PM2.5"

beta
parameter \(\beta\) (default is 2)

...
arguments to be passed to RMS_U_obs in MPI_bias and MPI_*_time, to U_obs_95_year in MPI_*_space, to U_obs_95 in MPI_perc

Value

MPI_bias: Modelling Performance Indicator for the bias, as in eq.26, tab.4, p.27 in Janssen et al., 2017 $$MPI_{bias} = \frac{BIAS}{\beta RMS_U}$$ MPI_corr_time: temporal Modelling Performance Indicator for the correlation, as in eq.27, tab.4, p.27 in Janssen et al., 2017 $$MPI_{corr,time} = \frac{1} { \left(R + \frac{\beta^2 RMS_U^2} { 2 \sigma_O \sigma_M} \right) }$$ MPI_sdev_time: temporal Modelling Performance Indicator for the standard deviation, as in eq.28, tab.4, p.27 in Janssen et al., 2017 $$MPI_{s.d.,time} = \frac{\sigma_M - \sigma_O}{\beta RMS_U}$$ MPI_corr_space: spatial Modelling Performance Indicator for the correlation, as in eq.29, tab.5, p.27 in Janssen et al., 2017 $$MPI_{corr,space} = \frac{1} { \left(R + \frac{\beta^2 RMS_{\bar{U}}^2} { 2 \sigma_{\bar{O}} \sigma_{\bar{M}}} \right) }$$ MPI_sdev_space: spatial Modelling Performance Indicator for the standard deviation, as in eq.30, tab.5, p.27 in Janssen et al., 2017 $$MPI_{s.d.,space} = \frac{\sigma_{\bar{M}} - \sigma_{\bar{O}}}{\beta RMS_{\bar{U}}}$$ MPI_perc: Modelling Performance Indicator for high percentile values, as in eq.35, p.35 in Janssen et al., 2017 $$MPI_{perc} = \frac{M_{perc} - O_{perc}}{\beta U_{95} (O_{perc})}$$

References

Janssen et al., 2017. "Guidance Document on Modelling Quality Objectives and Benchmarking. Version 2.1"