Modelling Quality Indicators

MQI_ts: MQI for a time series (vector)

MQI_ts(obs, mod, pollutant, beta = 2, ...)

MQI_ts_synth(obs, mod, pollutant, beta = 2, ...)

MQI_year(obs, mod, pollutant, beta = 2, ...)

Arguments

obs
numeric vector of observed values (yearly averages in MQI_year)

mod
numeric vector of modelled values (yearly averages in MQI_year)

pollutant
one of "NO2", "O3", "PM10", "PM2.5"

beta
parameter \(\beta\) (default is 2)

...
arguments to be passed to U_obs_95 in MQI_ts and MQI_ts_synth, to U_obs_95_year in MQI_year

Value

MQI_ts: Modelling Quality Indicator for a time series, as in eq.15, p.22 in Janssen et al., 2017 $$MQI = \frac{|O_i - M_i|}{\beta U_{95}(O_i)}$$

MQI_ts_synth: scalar synthetic Modelling Quality Indicator for a time series, as in eq.17, p.23 in Janssen et al., 2017 $$MQI = \frac{\sqrt{\frac{1}{N} \sum_{i=1}^N{(O_i-M_i)^2}}}{\beta \sqrt{\frac{1}{N} \sum_{i=1}^N{U_{95}(O_i)^2}}}$$

MQI_year: Modelling Quality Indicator for yearly average, as in eq.18, p.23 in Janssen et al., 2017 $$MQI = \frac{|\bar{O}-\bar{M}|}{\beta U_{95}(\bar{O})}$$

References

Janssen et al., 2017. "Guidance Document on Modelling Quality Objectives and Benchmarking. Version 2.1"